// src/pages/Board/react-beautiful-dnd.d.ts
export interface DraggableLocation {
  droppableId: DroppableId;
  index: number;
}
// ...
export interface Combine {
  draggableId: DraggableId;
  droppableId: DroppableId;
}
// ...
export interface DragStart extends BeforeCapture {
  type: TypeId;
  source: DraggableLocation;
}

// Published in onBeforeCapture
// We cannot give more information as things might change in the
// onBeforeCapture responder!
export interface BeforeCapture {
  draggableId: DraggableId;
  mode: MovementMode;
}

// published when a drag starts
export interface DragStart extends DraggableRubric {
  mode: MovementMode;
}

export interface DragUpdate extends DragStart {
  // may not have any destination (drag to nowhere)
  destination?: DraggableLocation | undefined;
  // populated when a draggable is dragging over another in combine mode
  combine?: Combine | undefined;
}

export type DropReason = 'DROP' | 'CANCEL';

export interface DropResult extends DragUpdate {
  reason: DropReason;
}